Update generic_param.h for fixing another warnings
New warnings appeared after fixing the previous one

+++ source/operator/prototype/generic_param.h
@@ -28,8 +28,8 @@
 struct generic_param
 {
     const char* op_name; // what real action?
-    int max_input_num;
-    int max_output_num;
+    size_t max_input_num;
+    size_t max_output_num;
 };
 
 #endif
